Right align fix on artwork page.
It is happening on all browsers and it is solely because the Helmar
Pharoh’s Choice Cabinet  <li> is taller than the other <li>, because
the title is two lines instead of one like the others.

To fix it I added min-height: 2.2em to the .work-list li .title class.
Since the line height is set to 1.1em by setting the height to 2.2em I
guaranteed that all the titles would be at lease 2 lines tall. This
made them uniform.